Now enables relay on digital pin 12
Also moved the date printing to a single function to clean the code up a bit. For testing purposes the relay is triggered for 3 seconds on startup if there is still time remaining on the last charge but then will turn off until a new card is scanned. Still no card validation.